feat: 수업 추가에 memberid 전달 (ing)

+++ app/src/main/java/com/dogandpigs/fitnote/presentation/lesson/memberlessonlist/MemberLessonListNavigation.kt
@@ -9,7 +9,7 @@ import com.dogandpigs.fitnote.presentation.navigation.ARGUMENT_MEMBER_ID
 fun NavGraphBuilder.addMemberLessonList(
     route: String,
     popBackStack: () -> Unit,
-    navigateToAddLesson: () -> Unit,
+    navigateToAddLesson: (Int) -> Unit,
 ) {
     val argument = ARGUMENT_MEMBER_ID
